# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api to update and create relations for items and more functions",
        "version": "0.1"
    },
    "paths": {
...
        "/demonstrativo/conta": {
            "post": {
                "tags": [
                    "Demonstrativo"
                ],
                "summary": "Cria uma nova conta no demonstrativo",
                "description": "Cria uma nova conta no demonstrativo no sistema",
                "operationId": "136675d98ff653d7289d72a0e6b87f07",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "linha": {
                                        "description": "Linha",
                                        "type": "string"
                                    },
                                    "nome_conta": {
                                        "description": "Nome da conta",
                                        "type": "string"
                                    },
                                    "empresa": {
                                        "description": "Empresa",
                                        "type": "integer"
                                    },
                                    "revenda": {
                                        "description": "Revenda",
                                        "type": "integer"
                                    },
                                    "client": {
                                        "description": "Cliente",
                                        "type": "integer"
                                    },
                                    "origem": {
                                        "description": "Origem",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Conta criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensagem de sucesso",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Mensagem de erro",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Linha não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Mensagem de erro",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/demonstrativo/conta/{linha}": {
            "put": {
                "tags": [
                    "Demonstrativo"
                ],
                "summary": "Atualiza uma conta no demonstrativo",
                "description": "Atualiza uma conta no demonstrativo no sistema",
                "operationId": "a7bef1f643c82d28d4d0d4299a4d5815",
                "parameters": [
                    {
                        "name": "linha",
                        "in": "path",
                        "description": "ID da linha",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome_conta": {
                                        "description": "Nome da conta",
                                        "type": "string"
                                    },
                                    "empresa": {
                                        "description": "Empresa",
                                        "type": "integer"
                                    },
                                    "revenda": {
                                        "description": "Revenda",
                                        "type": "integer"
                                    },
                                    "client": {
                                        "description": "Cliente",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Conta atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensagem de sucesso",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/demonstrativo/linha": {
            "post": {
                "tags": [
                    "Demonstrativo"
                ],
                "summary": "Cria uma nova linha no demonstrativo",
                "description": "Cria uma nova linha no demonstrativo no sistema",
                "operationId": "658463db0e0559e564279fbeafb24b8c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome_linha": {
                                        "description": "Nome da linha",
                                        "type": "string"
                                    },
                                    "tipo_linha": {
                                        "description": "Tipo da linha",
                                        "type": "string"
                                    },
                                    "empresa": {
                                        "description": "Empresa",
                                        "type": "integer"
                                    },
                                    "revenda": {
                                        "description": "Revenda",
                                        "type": "integer"
                                    },
                                    "client": {
                                        "description": "Cliente",
                                        "type": "integer"
                                    },
                                    "origem": {
                                        "description": "Origem",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Linha criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensagem de sucesso",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Erros de validação",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/demonstrativo/linhas": {
            "post": {
                "tags": [
                    "Demonstrativo"
                ],
                "summary": "Retorna linhas do demonstrativo por cliente e empresa",
                "description": "Retorna linhas do demonstrativo por cliente e empresa",
                "operationId": "081b8ad3dcd165b7f7c3f4673f6e4aff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client": {
                                        "description": "ID do cliente",
                                        "type": "integer"
                                    },
                                    "empresa": {
                                        "description": "ID da empresa",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Linhas encontradas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "linha": {
                                            "description": "ID da linha",
                                            "type": "integer"
                                        },
                                        "nome_linha": {
                                            "description": "Nome da linha",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}